ReviewAdviceLetter reusable component
